refactor: slight changes
