+++ src/shared/icons/index.ts
@@ -7,6 +7,7 @@ import { FacebookIcon } from './facebook-icon';
 import { TelegramIcon } from './telegram-icon';
 import { CloseMenuIcon } from './close-menu-icon';
 import { BreadcrumbsArrowIcon } from './breadcrumbs-arrow-icon';
+import { ResetInputIcon } from './reset-input-icon';
 
 export const Icons = {
   LocationIcon,
@@ -18,4 +19,5 @@ export const Icons = {
   TelegramIcon,
   CloseMenuIcon,
   BreadcrumbsArrowIcon,
+  ResetInputIcon,
 };
